fix guildmemberremove 2

+++ src/events/GuildMemberRemove.ts
@@ -7,33 +7,38 @@ export default class GuildMemberRemove implements Event {
     once = false;
 
     execute(client: BotClient, member: GuildMember): void {
-        member.ban({ reason: 'Покинул сервер' }).catch(error => client.logger.error(error));
-        client.db.userPoll
-            .findFirst({
-                where: { memberId: member.id },
-            })
-            .then(memberDb => {
-                if (memberDb) {
-                    client.db.userPoll
-                        .delete({
-                            where: { memberId: member.id },
-                        })
-                        .catch(error => client.logger.error(error));
-                    const voteChannel = client.channels.cache.get(client.config.voteChannel);
-                    if (voteChannel && voteChannel instanceof TextChannel) {
-                        voteChannel.messages
-                            .fetch(memberDb.messageId)
-                            .then(message => message.delete())
+        if (!member.roles.cache.get(client.config.role)) {
+            member.ban({ reason: 'Вышел во время голосования' }).catch(error => client.logger.error(error));
+
+            client.db.userPoll
+                .findFirst({
+                    where: { memberId: member.id },
+                })
+                .then(memberDb => {
+                    if (memberDb) {
+                        client.db.userPoll
+                            .delete({
+                                where: { memberId: member.id },
+                            })
                             .catch(error => client.logger.error(error));
+                        const voteChannel = client.channels.cache.get(client.config.voteChannel);
+                        if (voteChannel && voteChannel instanceof TextChannel) {
+                            voteChannel.messages
+                                .fetch(memberDb.messageId)
+                                .then(message => message.delete())
+                                .catch(error => client.logger.error(error));
+                        }
                     }
-                }
-            })
-            .catch(error => client.logger.error(error));
+                })
+                .catch(error => client.logger.error(error));
 
-        const bannedChannel = client.channels.cache.get(client.config.bannedChannel) as TextChannel;
+            const bannedChannel = client.channels.cache.get(client.config.bannedChannel) as TextChannel;
 
-        bannedChannel
-            .send(`${member.displayName} (${member.user.tag}) забанен. Причина: **покинул сервер**.`)
-            .catch(error => client.logger.error(error));
+            bannedChannel
+                .send(
+                    `${member.displayName} (${member.user.tag}) забанен. Причина: **Вышел во время голосования**.`,
+                )
+                .catch(error => client.logger.error(error));
+        }
     }
 }
